remove other methods from signalement repository  #5622
