+++ src/Service/Statistics/LogementDesordresStatisticProvider.php
@@ -4,11 +4,12 @@ namespace App\Service\Statistics;
 
 use App\Entity\Enum\DesordreCritereZone;
 use App\Entity\Territory;
-use App\Repository\SignalementRepository;
+use App\Repository\Query\Statistics\CountStatisticsQuery;
 
 class LogementDesordresStatisticProvider
 {
-    public function __construct(private SignalementRepository $signalementRepository)
+    public function __construct(
+        private CountStatisticsQuery $countStatisticsQuery, )
     {
     }
 
@@ -17,7 +18,7 @@ class LogementDesordresStatisticProvider
      */
     public function getData(?Territory $territory, ?int $year): array
     {
-        $countPerLogementDesordres = $this->signalementRepository->countByDesordresCriteres(
+        $countPerLogementDesordres = $this->countStatisticsQuery->countByDesordresCriteres(
             $territory,
             $year,
             DesordreCritereZone::LOGEMENT
