+++ tests/Functional/Service/Statistics/BatimentDesordresStatisticProviderTest.php
@@ -2,7 +2,7 @@
 
 namespace App\Tests\Functional\Service\Statistics;
 
-use App\Repository\SignalementRepository;
+use App\Repository\Query\Statistics\CountStatisticsQuery;
 use App\Service\Statistics\BatimentDesordresStatisticProvider;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 
@@ -11,9 +11,9 @@ class BatimentDesordresStatisticProviderTest extends KernelTestCase
     public function testGetData(): void
     {
         self::bootKernel();
-        /** @var SignalementRepository $signalementRepository */
-        $signalementRepository = self::getContainer()->get(SignalementRepository::class);
-        $data = (new BatimentDesordresStatisticProvider($signalementRepository))->getData(null, null);
+        /** @var CountStatisticsQuery $countStatisticsQuery */
+        $countStatisticsQuery = self::getContainer()->get(CountStatisticsQuery::class);
+        $data = (new BatimentDesordresStatisticProvider($countStatisticsQuery))->getData(null, null);
 
         $this->assertEquals(5, \count($data));
         $this->assertArrayHasKey('label', $data[0]);
